feat: add mqtt5.0

+++ mqtt-broker/src/main/java/org/github/zzf/mqtt/protocol/server/Authenticator.java
@@ -11,8 +11,9 @@ public interface Authenticator {
     int AUTHENTICATE_SUCCESS = 0x00;
 
     /**
-     * @return 0x00 authenticate success; 0x02 Connection Refused, identifier rejected; 0x04 Connection Refused, bad username or password
+     * @return 0x00 authenticate success; 0x02 Connection Refused, identifier rejected; 0x04 Connection Refused, bad
+     * username or password
      */
-    int authenticate(Connect packet);
+    byte authenticate(Connect packet);
 
 }
